Initial commit of ugly branch.
