--- src/gburtini/Distributions/IDistribution.php
@@ -1,12 +1,0 @@
-<?php 
-	namespace gburtini\Distributions;
-
-	interface IDistribution {
-		public function pdf($x);
-		public function pmf($x);
-		public function cdf($x);
-		public function icdf($y);
-		public function quantile($y);
-		public function rand();
-	}
-?>
