+++ src/gburtini/Distributions/T.php
@@ -4,36 +4,11 @@
 	 *
 	 * Copyright (C) 2015 Giuseppe Burtini <joe@iterative.ca> except where otherwise noted.
 	 *
-     * Other Credits
-     * -------------
-     * Some work derived (with permission/license) from jStat - Javascript Statistical Library (MIT licensed).
+	 * Other Credits
+	 * -------------
+	 * Some work derived (with permission/license) from jStat - Javascript Statistical Library (MIT licensed).
 	 */
 
 	namespace gburtini\Distributions;
-	use gburtini\Distributions\Distribution;
-	use gburtini\Distributions\Accessories\BetaFunction;
-
-	class T extends Distribution {
-		protected $degrees;
-		public function __construct($dof) {
-			$this->degrees = floatval($dof); // float, not integer: http://stats.stackexchange.com/questions/116511/explanation-for-non-integer-degrees-of-freedom-in-t-test-with-unequal-variances
-			// TODO: some overflow problems to be dealt with here (large DOF)
-		}
-	
-		public function pdf($x) {
-			return (1 / (sqrt($this->degrees) * BetaFunction::betaFunction(0.5, $this->degrees/2))) * pow(1 + (($x*$x)/$this->degrees), -(($this->degrees + 1) / 2));
-  		}
-
-		public function cdf($x) {
-			$halfDegrees = $this->degrees/2;
-			return BetaFunction::incompleteBetaFunction( ($x + sqrt($x*$x + $this->degrees)) / (2 * sqrt($x*$x + $this->degrees)), $halfDegrees, $halfDegrees);
-		}
-
-		public function icdf($y) {
-			$x = BetaFunction::inverseIncompleteBetaFunction(2 * min($y, 1 - $y), 0.5 * $this->degrees, 0.5);
-			$x = sqrt($this->degrees * (1 - $y) / $y);
-			return ($y > 0.5) ? $x : -$x;
-		}
-	}
-
-?>
+	require_once dirname(__FILE__) . "/../../../ugly/T.php";
+	class T extends GBPDP_T {}
